requirements.txt
# ─── Core web framework ───────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# ─── Telegram bot ─────────────────────────────────────────────────────────────
aiogram>=3.10.0
aiohttp>=3.9.5
aiofiles>=23.2.1

# ─── AI / LLM ─────────────────────────────────────────────────────────────────
openai>=1.30.0

# ─── Google APIs ──────────────────────────────────────────────────────────────
google-api-python-client>=2.130.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# ─── Database ─────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.30
# Uncomment for PostgreSQL support:
# psycopg2-binary>=2.9.9

# ─── HTTP client (OAuth callbacks) ────────────────────────────────────────────
httpx>=0.27.0

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest>=8.2.0
pytest-asyncio>=0.23.7
httpx>=0.27.0
